pyproject.toml
[tool.poetry]
name = "order-service"
version = "0.1.0"
description = ""
authors = ["Denis <den@listapad@mail.ru>"]
readme = "README.md"
packages = [{include = "order_service"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2 = "^2.9.9"
asyncpg = "^0.29.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.3"
httpx = "^0.26.0"
aiohttp = "^3.9.1"
jinja2 = "^3.1.2"
websockets = "^12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
